user menu refactor
