+++ components/navbar/user.tsx
@@ -1,0 +1,76 @@
+import { signOut, useSession } from 'next-auth/react'
+import { useRouter } from 'next/router'
+import styles from '../../styles/form.module.css'
+
+const UserMenuNavBar = ({ menu }: { menu: boolean }) => {
+  const { data: session } = useSession()
+  const router = useRouter()
+
+  return (
+    <>
+      {session ? (
+        <div className={`relative ${menu ? '' : 'hidden'} ${styles.menuLogIn}`}>
+          <div
+            className={` z-[999] w-44 shadow bg-gray-700 divide-gray-600 absolute top-9 right-5 rounded-xl `}
+          >
+            <div className="py-3 px-4 text-sm text-white hover:bg-gray-600 hover:rounded-t-xl">
+              <div className="font-medium truncate">{session?.user.email}</div>
+            </div>
+            <ul
+              className="py-1 text-sm text-gray-200 border-y-white border-y-2"
+              aria-labelledby="dropdownInformationButton"
+            >
+              <li className="block py-2 px-4 hover:bg-gray-600 hover:text-white">
+                <a href="#">Dashboard</a>
+              </li>
+              <li className="block py-2 px-4 hover:bg-gray-600 hover:text-white">
+                <a href="#">Settings</a>
+              </li>
+            </ul>
+            <div
+              className="block py-3 px-4 text-sm hover:bg-gray-600 hover:rounded-b-xl text-gray-200 hover:text-white"
+              onClick={async () => {
+                await signOut()
+              }}
+            >
+              <a href="#">Sign out</a>
+            </div>
+          </div>
+        </div>
+      ) : (
+        <div
+          className={`relative ${menu ? '' : 'hidden'} ${styles.menuLogOut}`}
+        >
+          <div
+            className={` z-[999] w-44 shadow bg-gray-700 absolute top-9 right-[-2rem] rounded-xl `}
+          >
+            <div className="py-3 px-4 text-sm text-white">Welcome user !</div>
+            <ul
+              className=" text-sm text-gray-200 border-t-2 border-gray-600"
+              aria-labelledby="dropdownInformationButton"
+            >
+              <li
+                className="block py-3 px-4 hover:bg-gray-600 hover:text-white cursor-pointer"
+                onClick={() => {
+                  router.push('/logIn')
+                }}
+              >
+                LogIn
+              </li>
+              <li
+                className="block py-3 pb-4 px-4 hover:bg-gray-600 hover:text-white hover:rounded-b-xl cursor-pointer"
+                onClick={() => {
+                  router.push('/signIn')
+                }}
+              >
+                Register
+              </li>
+            </ul>
+          </div>
+        </div>
+      )}
+    </>
+  )
+}
+
+export default UserMenuNavBar
