+++ styles/form.module.css
@@ -57,11 +57,21 @@
   }
 }
 
-.menu::before {
+.menuLogIn::before {
   position: absolute;
   content: '';
   top: 1.3rem;
   right: 5rem;
+  border-bottom: 1rem solid rgb(55, 65, 81, 70);
+  border-left: 1.5rem solid transparent;
+  border-right: 1.5rem solid transparent;
+  z-index: 20;
+}
+.menuLogOut::before {
+  position: absolute;
+  content: '';
+  top: 1.5rem;
+  right: 0;
   border-bottom: 1rem solid rgb(55, 65, 81, 70);
   border-left: 1.5rem solid transparent;
   border-right: 1.5rem solid transparent;
